requirements.txt
aniso8601==3.0.0
blinker==1.4
bson==0.5.7
certifi==2018.10.15
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Compress==1.4.0
Flask-Cors==3.0.4
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
pkg-resources==0.0.0
pprint==0.1
psycopg2-binary==2.7.5
PyJWT==1.4.2
pymongo==3.7.1
python-dateutil==2.7.5
pytz==2018.4
requests==2.20.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
